+++ Rebus/Pipeline/Send/AutoHeadersOutgoingStep.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using Rebus.Messages;
+// ReSharper disable ForCanBeConvertedToForeach
 
 namespace Rebus.Pipeline.Send
 {
@@ -17,7 +18,7 @@ namespace Rebus.Pipeline.Send
 Headers already on the message will not be overwritten.")]
     public class AutoHeadersOutgoingStep : IOutgoingStep
     {
-        readonly ConcurrentDictionary<Type, IEnumerable<HeaderAttribute>> _headersToAssign = new ConcurrentDictionary<Type, IEnumerable<HeaderAttribute>>();
+        readonly ConcurrentDictionary<Type, HeaderAttribute[]> _headersToAssign = new ConcurrentDictionary<Type, HeaderAttribute[]>();
 
         /// <summary>
         /// Carries out the auto-header logic
@@ -35,10 +36,12 @@ Headers already on the message will not be overwritten.")]
                 .GetTypeInfo()
                 .GetCustomAttributes(typeof (HeaderAttribute), true)
                 .OfType<HeaderAttribute>()
-                .ToList());
+                .ToArray());
 
-            foreach (var autoHeader in headersToAssign)
+            for (var index = 0; index < headersToAssign.Length; index++)
             {
+                var autoHeader = headersToAssign[index];
+
                 if (headers.ContainsKey(autoHeader.Key)) continue;
 
                 headers[autoHeader.Key] = autoHeader.Value;